fix register and some bugs and also learn to use postman

+++ src/utils/cloudinary.js
@@ -1,12 +1,14 @@
+import dotenv from 'dotenv';
+dotenv.config();
 import { v2 as cloudinary } from 'cloudinary';
 import fs from "fs";
 
 
     // Configuration
     cloudinary.config({ 
-        cloud_name: 'process.env.CLOUDINARY_CLOUD_NAME', 
-        api_key: 'process.env.CLOUDINARY_API_KEY', 
-        api_secret: 'process.env.CLOUDINARY_API_SECRET' // Click 'View API Keys' above to copy your API secret
+        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+        api_key: process.env.CLOUDINARY_API_KEY, 
+        api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
     });
 
 
@@ -19,12 +21,16 @@ import fs from "fs";
                 resource_type: "auto"
             })
             // file has been uploaded successfully
-            console.log("File is uploaded on cloudinary", 
-                response.url);
+            // console.log("File is uploaded on cloudinary", 
+            //     response.url);
+            fs.unlinkSync(localFilePath)
                 return response;
             
         } catch (error) {
-            fs.unlinkSync(localFilePath) // remove the file from local storage
+            console.log("Cloudinary Error", error);
+           if (fs.existsSync(localFilePath)) {
+                fs.unlinkSync(localFilePath)
+            }
             return null
         }
       }
